move old contribution to legacy module
