+++ lib/inss_calculator.rb
@@ -7,13 +7,13 @@ require_relative 'inss_calculator/discount_calculator_base'
 require_relative 'inss_calculator/discount_previdence_calculator'
 require_relative 'inss_calculator/public_inss_calculator'
 require_relative 'inss_calculator/legacy/first_discount_calculator'
-require_relative 'inss_calculator/second_discount_calculator'
-require_relative 'inss_calculator/third_discount_calculator'
-require_relative 'inss_calculator/fourth_discount_calculator'
-require_relative 'inss_calculator/fifth_discount_calculator'
-require_relative 'inss_calculator/sixth_discount_calculator'
-require_relative 'inss_calculator/seventh_discount_calculator'
-require_relative 'inss_calculator/eigth_discount_calculator'
+require_relative 'inss_calculator/legacy/second_discount_calculator'
+require_relative 'inss_calculator/legacy/third_discount_calculator'
+require_relative 'inss_calculator/legacy/fourth_discount_calculator'
+require_relative 'inss_calculator/legacy/fifth_discount_calculator'
+require_relative 'inss_calculator/legacy/sixth_discount_calculator'
+require_relative 'inss_calculator/legacy/seventh_discount_calculator'
+require_relative 'inss_calculator/legacy/eigth_discount_calculator'
 require_relative 'inss_calculator/decorator/text'
 
 module InssCalculator
