+++ lib/inss_calculator/legacy/salary_levels.rb
@@ -1,29 +1,27 @@
 module InssCalculator
   module Legacy
-    module SalaryLevels
-      FIRST_SALARY_BASE = 0.0
-      FIRST_SALARY_LIMIT = 1412.0
+    FIRST_SALARY_BASE = 0.0
+    FIRST_SALARY_LIMIT = 1412.0
 
-      SECOND_SALARY_BASE = 1412.01
-      SECOND_SALARY_LIMIT = 2666.68
+    SECOND_SALARY_BASE = 1412.01
+    SECOND_SALARY_LIMIT = 2666.68
 
-      THIRD_SALARY_BASE = 2666.69
-      THIRD_SALARY_LIMIT = 4000.03
+    THIRD_SALARY_BASE = 2666.69
+    THIRD_SALARY_LIMIT = 4000.03
 
-      FOURTH_SALARY_BASE = 4000.04
-      FOURTH_SALARY_LIMIT = 7786.02
+    FOURTH_SALARY_BASE = 4000.04
+    FOURTH_SALARY_LIMIT = 7786.02
 
-      FIFTH_SALARY_BASE = 7786.03
-      FIFTH_SALARY_LIMIT = 13_333.48
+    FIFTH_SALARY_BASE = 7786.03
+    FIFTH_SALARY_LIMIT = 13_333.48
 
-      SIXTH_SALARY_BASE = 13_333.49
-      SIXTH_SALARY_LIMIT = 26_666.94
+    SIXTH_SALARY_BASE = 13_333.49
+    SIXTH_SALARY_LIMIT = 26_666.94
 
-      SEVENTH_SALARY_BASE = 26_666.95
-      SEVENTH_SALARY_LIMIT = 52_000.54
+    SEVENTH_SALARY_BASE = 26_666.95
+    SEVENTH_SALARY_LIMIT = 52_000.54
 
-      EIGTH_SALARY_BASE = 52_000.55
-      EIGTH_SALARY_LIMIT = Float::INFINITY
-    end
+    EIGTH_SALARY_BASE = 52_000.55
+    EIGTH_SALARY_LIMIT = Float::INFINITY
   end
 end
