+++ spec/inss_calculator/legacy/third_discount_calculator_spec.rb
@@ -4,7 +4,7 @@ require 'spec_helper'
 
 include InssCalculator
 
-RSpec.describe ThirdDiscountCalculator do
+RSpec.describe Legacy::ThirdDiscountCalculator do
   it 'is true' do
     expect(described_class.new('3000')).to be_truthy
   end
